+++ SWP391_Group6/nbproject/private/private.properties
@@ -1,5 +1,10 @@
-deploy.ant.properties.file=C:\\Users\\Admin\\AppData\\Roaming\\NetBeans\\17\\tomcat101.properties
+deploy.ant.properties.file=C:\\Users\\ADMIN\\AppData\\Roaming\\NetBeans\\17\\tomcat101.properties
+file.reference.jakarta.activation-2.0.1.jar=E:\\SWP391_HappyPrograming-main\\SWP391_Group6\\lib\\lib\\jakarta.activation-2.0.1.jar
+file.reference.jakarta.mail-2.0.1.jar=E:\\SWP391_HappyPrograming-main\\SWP391_Group6\\lib\\lib\\jakarta.mail-2.0.1.jar
+file.reference.jakarta.servlet.jsp.jstl-2.0.0.jar=E:\\SWP391_HappyPrograming-main\\SWP391_Group6\\lib\\lib\\jakarta.servlet.jsp.jstl-2.0.0.jar
+file.reference.jakarta.servlet.jsp.jstl-api-2.0.0.jar=E:\\SWP391_HappyPrograming-main\\SWP391_Group6\\lib\\lib\\jakarta.servlet.jsp.jstl-api-2.0.0.jar
+file.reference.mssql-jdbc-12.4.2.jre11.jar=E:\\SWP391_HappyPrograming-main\\SWP391_Group6\\lib\\lib\\mssql-jdbc-12.4.2.jre11.jar
 j2ee.server.home=C:/Program Files/Apache Software Foundation/Tomcat 10.1
 j2ee.server.instance=tomcat101:home=C:\\Program Files\\Apache Software Foundation\\Tomcat 10.1
 selected.browser=Chrome.INTEGRATED
-user.properties.file=C:\\Users\\Admin\\AppData\\Roaming\\NetBeans\\17\\build.properties
+user.properties.file=C:\\Users\\ADMIN\\AppData\\Roaming\\NetBeans\\17\\build.properties
