+++ SWP391_Group6/nbproject/private/private.xml
@@ -2,6 +2,14 @@
 <project-private xmlns="http://www.netbeans.org/ns/project-private/1">
     <editor-bookmarks xmlns="http://www.netbeans.org/ns/editor-bookmarks/2" lastBookmarkId="0"/>
     <open-files xmlns="http://www.netbeans.org/ns/projectui-open-files/2">
-        <group/>
+        <group>
+            <file>file:/E:/SWP391_HappyPrograming-main/SWP391_Group6/src/java/controller/ChangePasswordController.java</file>
+            <file>file:/E:/SWP391_HappyPrograming-main/SWP391_Group6/src/java/controller/LogoutController.java</file>
+            <file>file:/E:/SWP391_HappyPrograming-main/SWP391_Group6/src/java/dal/AccountDBContext.java</file>
+            <file>file:/E:/SWP391_HappyPrograming-main/SWP391_Group6/src/java/dal/DBContext.java</file>
+            <file>file:/E:/SWP391_HappyPrograming-main/SWP391_Group6/web/WEB-INF/view/user/login.jsp</file>
+            <file>file:/E:/SWP391_HappyPrograming-main/SWP391_Group6/src/java/controller/HomeController.java</file>
+            <file>file:/E:/SWP391_HappyPrograming-main/SWP391_Group6/src/java/controller/LoginController.java</file>
+        </group>
     </open-files>
 </project-private>
