+++ SWP391_Group6/src/java/dal/AccountDBContext.java
@@ -36,13 +36,14 @@ public class AccountDBContext extends DBContext<Account> {
     public void insert(Account entity) {
         try {
             connection.setAutoCommit(false);
-            String sql = "INSERT INTO Account VALUES(?, ?, ?, ?, ?)";
+            String sql = "INSERT INTO Account VALUES(?, ?, ?, ?, ?, ?)";
             PreparedStatement ps = connection.prepareStatement(sql);
-            ps.setString(1, entity.getUsername());
-            ps.setString(2, entity.getPassword());
-            ps.setString(3, entity.getRole());
-            ps.setString(4, entity.getEmail());
-            ps.setBoolean(5, entity.isStatus());
+            ps.setInt(1, entity.getId());
+            ps.setString(2, entity.getUsername());
+            ps.setString(3, entity.getPassword());
+            ps.setString(4, entity.getRole());
+            ps.setString(5, entity.getEmail());
+            ps.setBoolean(6, entity.isStatus());
 
             ps.executeUpdate();
             connection.commit();
