+++ SWP391_Group6/src/java/dal/DBContext.java
@@ -7,7 +7,9 @@ package dal;
 import controller.LoginController;
 import java.util.logging.Level;
 import java.util.logging.Logger;
-import java.sql.*;
+import java.sql.Connection;
+import java.sql.DriverManager;
+import java.sql.SQLException;
 import java.util.ArrayList;
 import model.User;
 
@@ -23,7 +25,7 @@ public abstract class DBContext<T> {
     public DBContext() {
         try {
             String user = "sa";
-            String pass = "sa123";
+            String pass = "123456";
             String url = "jdbc:sqlserver://localhost:1433;databaseName=HappyPrograming;encrypt=true;trustServerCertificate=true;";
             Class.forName("com.microsoft.sqlserver.jdbc.SQLServerDriver");
             connection = DriverManager.getConnection(url, user, pass);
