+++ SWP391_Group6/web/WEB-INF/web.xml
@@ -26,7 +26,7 @@
     </servlet>
     <servlet>
         <servlet-name>ForgotAccountController</servlet-name>
-        <servlet-class>controller.ForgotPassController</servlet-class>
+        <servlet-class>controller.ForgotAccountController</servlet-class>
     </servlet>
     <servlet>
         <servlet-name>SkillController</servlet-name>
